+++ src/fullflow/System/__init__.py
@@ -5,3 +5,4 @@ from .Components import *
 from .Network import *
 from .Composition import *
 from .Model import *
+from .protocols import *